scripts/update_configs_from_dataset.py
#!/usr/bin/env python3
"""
Update configuration files based on actual dataset analysis
Phase: 1 (after dataset analysis complete)
"""

import json
import yaml
import pandas as pd
from pathlib import Path

def analyze_dataset_patterns(cell_data, ue_data):
    """Analyze dataset to extract actual patterns."""
    
    patterns = {}
    
    # MIMO zero rate analysis
    if 'CARR.AverageLayersDl' in cell_data.columns:
        patterns['mimo_zero_rate'] = (cell_data['CARR.AverageLayersDl'] == 0).mean()
    
    # CQI zero rate analysis  
    if 'DRB.UECqiDl' in ue_data.columns:
        patterns['cqi_zero_rate'] = (ue_data['DRB.UECqiDl'] == 0).mean()
    
    # UL/DL symmetry analysis
    if len(cell_data) > 100:
        patterns['ul_dl_symmetry'] = cell_data['DRB.UEThpUl'].corr(cell_data['DRB.UEThpDl'])
    
    # Temporal pattern analysis
    cell_data['hour'] = pd.to_datetime(cell_data['timestamp']).dt.hour
    hourly_load = cell_data.groupby('hour')['RRC.ConnMean'].mean()
    peak_threshold = hourly_load.quantile(0.70)
    patterns['peak_hours'] = hourly_load[hourly_load >= peak_threshold].index.tolist()
    
    return patterns

def update_configurations(patterns):
    """Update config files with derived patterns."""
    
    # Update drift thresholds
    config_path = Path('config/drift_thresholds.yaml')
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    
    # Update pattern recognition
    if 'peak_hours' in patterns:
        config['drift_detection']['pattern_recognition']['peak_hours'] = patterns['peak_hours']
    
    # Save updated config
    with open(config_path, 'w') as f:
        yaml.dump(config, f, default_flow_style=False)
    
    print(f"Updated {config_path} with dataset-derived patterns")

if __name__ == "__main__":
    print("This script will be called after dataset analysis in Phase 1")
