requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
scipy==1.10.1
pyyaml>=5.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
optuna==3.3.0
kafka-python==2.0.2
pyarrow==14.0.1
ricxappframe==3.2.1
msgpack==1.0.5
mdclogpy==1.1.4
